+++ src/components/topbar/index.js
@@ -1,0 +1,3 @@
+import main from './topbar';
+
+export default main;